requirements.txt
get==2019.4.13
Pillow==5.4.1
post==2019.4.13
public==2019.4.13
query-string==2019.4.13
request==2019.4.13
